Move common placementpolicy components to TopologyAwareEnsemblePlacementPolicy.
- Moving components/methods which are common in nature with respect to
placementpolicy from RackawareEnsemblePlacementPolicyImpl to
TopologyAwareEnsemblePlacementPolicy, so that any new placementpolicy
implementation can extend TopologyAwareEnsemblePlacementPolicy and reuse
those common components/generic.
- Change signature of methods of RackChangeNotifier interface.
- Delete duplicate methods in RackawareEnsemblePlacementPolicyImpl and
TopologyAwareEnsemblePlacementPolicy (missed removing them in #2089)
- Created separate class for BookieNode
- This change has no functionality change, it is just reorganizing code.
