+++ bookkeeper-server/src/main/java/org/apache/bookkeeper/net/BookieNode.java
@@ -1,0 +1,53 @@
+/**
+ * Licensed to the Apache Software Foundation (ASF) under one
+ * or more contributor license agreements.  See the NOTICE file
+ * distributed with this work for additional information
+ * regarding copyright ownership.  The ASF licenses this file
+ * to you under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance
+ * with the License.  You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package org.apache.bookkeeper.net;
+
+/**
+ * Bookie Node implementation.
+ */
+public class BookieNode extends NodeBase {
+    private final BookieSocketAddress addr; // identifier of a bookie node.
+
+    public BookieNode(BookieSocketAddress addr, String networkLoc) {
+        super(addr.toString(), networkLoc);
+        this.addr = addr;
+    }
+
+    public BookieSocketAddress getAddr() {
+        return addr;
+    }
+
+    @Override
+    public int hashCode() {
+        return name.hashCode();
+    }
+
+    @Override
+    public boolean equals(Object obj) {
+        if (!(obj instanceof BookieNode)) {
+            return false;
+        }
+        BookieNode other = (BookieNode) obj;
+        return getName().equals(other.getName());
+    }
+
+    @Override
+    public String toString() {
+        return String.format("<Bookie:%s>", name);
+    }
+}
